--- app/client/cypress/integration/Smoke_TestSuite/ClientSideTests/Comments/utils.js
@@ -1,12 +1,0 @@
-export function typeIntoDraftEditor(selector, text) {
-  cy.get(selector).then((input) => {
-    var textarea = input.get(0);
-    textarea.dispatchEvent(new Event("focus"));
-
-    var textEvent = document.createEvent("TextEvent");
-    textEvent.initTextEvent("textInput", true, true, null, text);
-    textarea.dispatchEvent(textEvent);
-
-    textarea.dispatchEvent(new Event("blur"));
-  });
-}
