+++ app/client/cypress/manual_TestSuite/CommentedScriptFiles/AForceMigration_Spec.ts
@@ -1,7 +1,7 @@
-const widgetsPage = require("../../../locators/Widgets.json");
-import homePage from "../../../locators/HomePage";
+// const widgetsPage = require("../../../locators/Widgets.json");
+// import homePage from "../../locators/HomePage";
 
-describe("Migration Validate", function () {
+// describe("Migration Validate", function () {
 //   it("1. Import application and Validate Migration on pageload", function () {
 //     // import application
 //     cy.get(homePage.homeIcon).click();
@@ -504,4 +504,4 @@ describe("Migration Validate", function () {
 //       .should("contain", "#FFC13D");
 //     cy.get(widgetsPage.selectedTextSize).should("have.text", "24px");
 //   });
-});
+// });
