+++ app/client/cypress/manual_TestSuite/CommentedScriptFiles/Table_Duplicate_ColumnName_spec.js
@@ -1,6 +1,6 @@
-const widgetsPage = require("../../../../locators/Widgets.json");
-const dsl = require("../../../../fixtures/tableNewDsl.json");
-const commonlocators = require("../../../../locators/commonlocators.json");
+const widgetsPage = require("../../locators/Widgets.json");
+const dsl = require("../../fixtures/tableNewDsl.json");
+const commonlocators = require("../../locators/commonlocators.json");
 
 describe("prevent duplicate column name in table", function() {
   before(() => {
